requirements-client.txt
pyopenssl
requests
click
